+++ src/fontra/client/renderer/welcome.html
@@ -15,11 +15,17 @@
       <div class="landing-item" id="fontra-icon-header">
         <img src="./images/fontra-icon.svg" width="100" height="100">
         <div>
-          Installing fontra server...
+          Initialising...
         </div>
       </div>
 
     </div>
 
   </body>
+  <script type="module">
+
+    import { startupLandingPage } from "./js/landing/landing.js";
+    startupLandingPage();
+
+  </script>
 </html>
